app/requirements.txt
anyio==3.6.2
arrow==1.2.3
bcrypt==4.0.1
Brotli==1.0.9
brotli-asgi==1.3.0
certifi==2022.12.7
click==8.1.3
cssselect==1.2.0
dnspython==2.3.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
motor==3.1.1
packaging==23.0
parsel==1.7.0
PyJWT==2.6.0
pymongo==4.3.3
python-dateutil==2.8.2
python-multipart==0.0.5
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.23.1
uvicorn==0.20.0
w3lib==2.1.1